+++ packages/website/src/demo-building-blocs/ExampleLayout.tsx
@@ -1,30 +1,46 @@
 import '@demo-styling/example-layout.scss';
 
-import {TabContent, TabPaneConnected, TabSelectors, TabsHeader, Tile} from '@coveord/plasma-react';
+import {SvgNames, TabContent, TabPaneConnected, TabSelectors, TabsHeader, Tile, TileProps} from '@coveord/plasma-react';
 import * as React from 'react';
 import {useSelector} from 'react-redux';
 
+import {GithubButton} from './GithubButton';
 import {GuidelinesTab} from './GuidelinesTab';
 import {PropsDoc} from './PropsDoc';
 import Sandbox from './Sandbox';
 
-export interface ExampleLayoutProps {
-    id: string;
+interface PlaygroundProps {
     title: string;
+    code: string;
+    layout?: 'horizontal' | 'vertical';
+}
+
+export interface ExampleLayoutProps extends PlaygroundProps {
+    id: string;
+    icon?: SvgNames;
     description?: React.ReactNode;
     section: string;
-    code: string;
-    examples: Record<string, {code: string; title: string}>;
+    examples?: Record<string, PlaygroundProps>;
+    relatedComponents?: TileProps[];
+    /**
+     * Path to the component's source file from /packages/react/src/components
+     *
+     * @example '/button/Button.tsx'
+     */
+    componentSourcePath: string;
 }
 
 export const ExampleLayout: React.FunctionComponent<ExampleLayoutProps> = ({
     id,
     title,
     description,
+    icon,
     section,
-    children,
     code,
+    layout = 'horizontal',
     examples,
+    componentSourcePath,
+    relatedComponents,
 }) => {
     const isShowingCode = useSelector((state) =>
         TabSelectors.getIsTabSelected(state, {groupId: 'page', id: 'implementation'})
@@ -33,9 +49,14 @@ export const ExampleLayout: React.FunctionComponent<ExampleLayoutProps> = ({
         <div id={id} className="example-layout">
             <div className="example-layout__header">
                 <h2 className="h4-subdued normal-white-space">{section}</h2>
-                <div>Github</div>
+                <GithubButton
+                    ariaLabel="View source code on GitHub"
+                    href={`https://github.com/coveo/plasma/blob/master/packages/react/src/components${componentSourcePath}`}
+                >
+                    View source
+                </GithubButton>
                 <h3 className="h1-light normal-white-space">{title}</h3>
-                <Tile title="" svgName="plasmaComponentBox" description="" href={undefined} />
+                <Tile svgName={icon || 'plasmaComponentBox'} />
                 <div>{description}</div>
             </div>
             <TabsHeader
@@ -46,37 +67,61 @@ export const ExampleLayout: React.FunctionComponent<ExampleLayoutProps> = ({
             />
             <TabContent>
                 <TabPaneConnected id="implementation" groupId="page">
-                    {isShowingCode && <Content id={id} code={code} examples={examples} />}
+                    {isShowingCode && (
+                        <Content
+                            id={id}
+                            code={code}
+                            examples={examples}
+                            relatedComponents={relatedComponents}
+                            layout={layout}
+                        />
+                    )}
                 </TabPaneConnected>
-                <GuidelinesTab id={id} />
+                <div className="mod-header-padding mod-form-top-bottom-padding">
+                    <GuidelinesTab id={id} />
+                </div>
             </TabContent>
         </div>
     );
 };
-const Content: React.FunctionComponent<Pick<ExampleLayoutProps, 'code' | 'examples' | 'id'>> = ({
-    code,
-    examples,
-    id,
-}) => (
+const Content: React.FunctionComponent<Pick<
+    ExampleLayoutProps,
+    'code' | 'examples' | 'id' | 'relatedComponents' | 'layout'
+>> = ({code, examples, id, relatedComponents, layout}) => (
     <>
         <div className="example-layout__main-code example-layout__section">
-            <Sandbox id="main-code">{code}</Sandbox>
+            <Sandbox id="main-code" horizontal={layout === 'horizontal'}>
+                {code}
+            </Sandbox>
         </div>
         <div className="example-layout__props example-layout__section">
             <h4 className="h2 mb1">Props</h4>
             <PropsDoc componentName={id} />
         </div>
-        <div className="example-layout__examples example-layout__section">
-            <h4 className="h2 mb5">Examples</h4>
-            {Object.entries(examples).map(([exampleId, {code: exampleCode, title}]) => (
-                <Sandbox key={exampleId} id={exampleId} title={title}>
-                    {exampleCode}
-                </Sandbox>
-            ))}
-        </div>
-        <div className="example-layout__related example-layout__section">
-            <h4 className="h2 mb5">Related Components</h4>
-            <div>Insert related components here</div>
-        </div>
+        {examples && (
+            <div className="example-layout__examples example-layout__section">
+                <h4 className="h2 mb5">Examples</h4>
+                {Object.entries(examples).map(
+                    ([exampleId, {code: exampleCode, title, layout: exampleLayout = 'horizontal'}]) => (
+                        <Sandbox
+                            key={exampleId}
+                            id={exampleId}
+                            title={title}
+                            horizontal={exampleLayout === 'horizontal'}
+                        >
+                            {exampleCode}
+                        </Sandbox>
+                    )
+                )}
+            </div>
+        )}
+        {relatedComponents && relatedComponents.length > 0 && (
+            <div className="example-layout__related example-layout__section">
+                <h4 className="h2 mb5">Related Components</h4>
+                {relatedComponents.map((tileProps) => (
+                    <Tile key={tileProps.title} {...tileProps} />
+                ))}
+            </div>
+        )}
     </>
 );